Move POJO classes to hbck-client package and adding Class description/comments
