+++ hbase-client/src/main/java/org/apache/hadoop/hbase/HbckRegionDetails.java
@@ -15,19 +15,40 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.hadoop.hbase.master.http.hbck.model;
+package org.apache.hadoop.hbase;
 
 import org.apache.yetus.audience.InterfaceAudience;
 
-@InterfaceAudience.Private
-public class HbckEmptyRegionInfo {
-  private final String regionInfo;
+/**
+ * POJO class for HBCK RegionInfo in HBCK Inconsistencies report.
+ */
+@InterfaceAudience.Public
+public class HbckRegionDetails {
+  private final String regionId;
+  private final String tableName;
+  private final String startKey;
+  private final String endKey;
 
-  public HbckEmptyRegionInfo(String emptyRegionInfo) {
-    this.regionInfo = emptyRegionInfo;
+  public HbckRegionDetails(String regionId, String tableName, String startKey, String endKey) {
+    this.regionId = regionId;
+    this.tableName = tableName;
+    this.startKey = startKey;
+    this.endKey = endKey;
   }
 
-  public String getRegionInfo() {
-    return regionInfo;
+  public String getRegionId() {
+    return regionId;
+  }
+
+  public String getTableName() {
+    return tableName;
+  }
+
+  public String getStartKey() {
+    return startKey;
+  }
+
+  public String getEndKey() {
+    return endKey;
   }
 }
